Moving Notifications repository methods to NotificationRespository
